management user

+++ src/main/java/com/an/tripora/repositories/DestinationRepo.java
@@ -8,6 +8,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DestinationRepo extends JpaRepository<Destination, Integer> {
+public interface DestinationRepo extends JpaRepository<Destination, Long> {
     Page<Destination> findByNameContainingIgnoreCase(String keyword, Pageable pageable);
 }
