+++ src/main/java/com/an/tripora/repositories/VehicleRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VehicleRepo extends JpaRepository<Vehicle, Integer> {
+public interface VehicleRepo extends JpaRepository<Vehicle, Long> {
 }
